Move `MaybeDyn` to new file and add some impls
