+++ packages/sycamore-reactive/src/utils.rs
@@ -4,9 +4,6 @@ use crate::*;
 
 /// A trait that is implemented for reactive data that can be accessed and tracked, such as
 /// [`Signal`].
-///
-/// You can make your components/functions take this instead of a concrete reactive type so that it
-/// can work with any reactive type.
 pub trait Accessor<T> {
     /// Get the reactive value. For example, with [`Signal`], this just calls
     /// [`get`](ReadSignal::get).
