+++ packages/sycamore-web/src/lib.rs
@@ -39,6 +39,7 @@ pub mod events;
 mod components;
 mod elements;
 mod iter;
+mod maybe_dyn;
 mod node;
 mod noderef;
 mod portal;
@@ -50,6 +51,7 @@ mod view;
 pub use self::components::*;
 pub use self::elements::*;
 pub use self::iter::*;
+pub use self::maybe_dyn::*;
 pub use self::node::*;
 pub use self::noderef::*;
 pub use self::portal::*;
