+++ packages/sycamore-web/src/node/mod.rs
@@ -99,74 +99,6 @@ impl AttributeValue for MaybeDynJsValue {
     }
 }
 
-/// Represents a value that can be either static or dynamic.
-pub enum MaybeDyn<T> {
-    Static(T),
-    Dynamic(Box<dyn FnMut() -> T>),
-}
-
-impl<T> MaybeDyn<T> {
-    /// Evaluate the value by consuming itself.
-    pub fn evaluate(self) -> T {
-        match self {
-            Self::Static(value) => value,
-            Self::Dynamic(mut f) => f(),
-        }
-    }
-}
-
-impl<T, F: FnMut() -> U + 'static, U: Into<T>> From<F> for MaybeDyn<T> {
-    fn from(mut f: F) -> Self {
-        Self::Dynamic(Box::new(move || f().into()))
-    }
-}
-
-/// A possibly dynamic string value.
-pub type MaybeDynString = MaybeDyn<Cow<'static, str>>;
-
-/// A possibly dynamic boolean value.
-pub type MaybeDynBool = MaybeDyn<bool>;
-
-/// A possibly dynamic [`JsValue`].
-pub type MaybeDynJsValue = MaybeDyn<JsValue>;
-
-macro_rules! impl_from_maybe_dyn {
-    ($struct:ty => $($ty:ty),*) => {
-        $(
-            impl From<$ty> for $struct {
-                fn from(value: $ty) -> Self {
-                    Self::Static(value.into())
-                }
-            }
-        )*
-    };
-}
-
-impl_from_maybe_dyn!(MaybeDynString => &'static str, String, Cow<'static, str>);
-
-impl_from_maybe_dyn!(MaybeDynBool => bool);
-
-impl_from_maybe_dyn!(
-    MaybeDynJsValue =>
-    JsValue,
-    String,
-    bool,
-    i8,
-    i16,
-    i32,
-    i64,
-    i128,
-    isize,
-    u8,
-    u16,
-    u32,
-    u64,
-    u128,
-    usize,
-    f32,
-    f64
-);
-
 thread_local! {
     /// Whether we are in hydration mode or not.
     pub(crate) static IS_HYDRATING: Cell<bool> = const { Cell::new(false) };
